mcp-oauth-dynamicclient/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-oauth-dynamicclient"
version = "0.1.0"
description = "MCP OAuth Dynamic Client Registration Server - OAuth 2.1 and RFC 7591 compliant"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "MCP OAuth Team"},
]
keywords = ["oauth", "oauth2", "mcp", "dynamic-client-registration", "rfc7591"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "httpx>=0.25.2",
    "redis>=5.0.1",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=41.0.7",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
mcp-oauth-server = "mcp_oauth_dynamicclient.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/mcp-oauth-dynamicclient"
Documentation = "https://mcp-oauth-dynamicclient.readthedocs.io"
Repository = "https://github.com/yourusername/mcp-oauth-dynamicclient"
Issues = "https://github.com/yourusername/mcp-oauth-dynamicclient/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mcp_oauth_dynamicclient = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/mcp_oauth_dynamicclient"]
omit = ["*/tests/*", "*/cli.py"]